Backed both graphs with same Node structure
